+++ Changelog.md
@@ -15,6 +15,7 @@
 - Fix: Vanishmod support does not work with the Protecto
 - Fix: Universal Block Reinforcers can be used under spawn protection
 - Fix: Several blocks (Key Panel, Passcode-protected Chest, Rift Stabilizer, Security Camera) don't keep their custom name when they're broken
+- Fix: Inventory Scanner dupe
 
 --------------------------Changelog for v1.9.8-beta1 of SecurityCraft--------------------------